# MEDRISE-MEDICAL-CENTRE-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedRise Medical Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing:
    get:
      operationId: listInvoices
      tags: [billing]
      summary: List invoices
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [billing]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /billing/{id}:
    get:
      operationId: getInvoice
      tags: [billing]
      summary: Get a single invoice with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    patch:
      operationId: updateInvoice
      tags: [billing]
      summary: Update invoice status or payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    delete:
      operationId: deleteInvoice
      tags: [billing]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /imaging/orders:
    get:
      operationId: listImagingOrders
      tags: [imaging]
      summary: List imaging orders
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: modality
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of imaging orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImagingOrder"
    post:
      operationId: createImagingOrder
      tags: [imaging]
      summary: Create a new imaging order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImagingOrderInput"
      responses:
        "201":
          description: Imaging order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagingOrder"
  /imaging/orders/{id}:
    patch:
      operationId: updateImagingOrder
      tags: [imaging]
      summary: Update an imaging order (status, findings, impression)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImagingOrderUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagingOrder"
        "404":
          description: Not found
    delete:
      operationId: deleteImagingOrder
      tags: [imaging]
      summary: Delete an imaging order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── SHIFTS & LEAVE ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    InvoiceItem:
      type: object
      required: [id, invoiceId, description, quantity, unitPrice, amount]
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: string
        amount:
          type: string
    Invoice:
      type: object
      required: [id, patientId, invoiceNumber, status, totalAmount, paidAmount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        consultationId:
          type: ["integer", "null"]
        invoiceNumber:
          type: string
        status:
          type: string
          enum: [unpaid, partial, paid, cancelled]
        totalAmount:
          type: string
        paidAmount:
          type: string
        paymentMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        createdAt:
          type: string
        updatedAt:
          type: string
    InvoiceInput:
      type: object
      required: [patientId, items]
      properties:
        patientId:
          type: integer
        consultationId:
          type: integer
        notes:
          type: string
        items:
          type: array
          items:
            type: object
            required: [description, quantity, unitPrice]
            properties:
              description:
                type: string
              quantity:
                type: integer
              unitPrice:
                type: number
    InvoiceUpdateInput:
      type: object
      properties:
        status:
          type: string
          enum: [unpaid, partial, paid, cancelled]
        paidAmount:
          type: number
        paymentMethod:
          type: string
        notes:
          type: string
...
    ImagingOrder:
      type: object
      required: [id, patientId, modality, priority, status, requestedAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        patientDob:
          type: ["string", "null"]
        consultationId:
          type: ["integer", "null"]
        queueEntryId:
          type: ["integer", "null"]
        requestedBy:
          type: ["integer", "null"]
        requestedByName:
          type: ["string", "null"]
        modality:
          type: string
        bodyPart:
          type: ["string", "null"]
        clinicalIndication:
          type: ["string", "null"]
        priority:
          type: string
          enum: [routine, urgent, stat]
        status:
          type: string
          enum: [requested, in-progress, completed, cancelled]
        findings:
          type: ["string", "null"]
        impression:
          type: ["string", "null"]
        reportedBy:
          type: ["integer", "null"]
        reportedByName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        requestedAt:
          type: string
        completedAt:
          type: ["string", "null"]
    ImagingOrderInput:
      type: object
      required: [patientId, modality]
      properties:
        patientId:
          type: integer
        consultationId:
          type: integer
        requestedBy:
          type: integer
        modality:
          type: string
        bodyPart:
          type: string
        clinicalIndication:
          type: string
        priority:
          type: string
          enum: [routine, urgent, stat]
        notes:
          type: string
    ImagingOrderUpdateInput:
      type: object
      properties:
        status:
          type: string
          enum: [requested, in-progress, completed, cancelled]
        findings:
          type: string
        impression:
          type: string
        reportedBy:
          type: integer
        notes:
          type: string

    # ── LAB SCHEMAS ───────────────────────────────────────────────────────
...